+++ controllers/passwordVault/deleteSavedPasswd.js
@@ -7,6 +7,12 @@ exports.deleteSavedPasswd = async (req, res) => {
   try {
     const verifyToken = req.user;
     const id = req.params.id;
+    // BOLA check: verify the password entry belongs to this user
+    const user = await User.findById(verifyToken.id);
+    if (!user.passwordVault.some((pwId) => pwId.toString() === id)) {
+      Response(res, false, "Not authorized to delete this password", 403);
+      return;
+    }
     try {
       await passwordvault.findByIdAndDelete(id);
       await User.findByIdAndUpdate(verifyToken.id, {
